PointLocators now use an axis mask instead of two different dimensions
This makes them more flexible, since we can say that only want the tree
to consider the x and z coordinate and ignore the y coordinate. This
change also enables the usecase for searching along only one axis (e.g.
to extract trench diameters at certain depths, by searching the closest
points along the vertical trench axis (point on the left sidewall and
point on the right sidewall) and use the distance between those two as a
measure for the diameter at that depth.