PointLocators now use an axis mask.
The axis mask allows selecting axes for nearest neighbors lookup. For
example: we could have a 4D data vector that consists of 3 spatial dimensions and 1 data dimension, so for the nearest neighbors lookup we would only want to use the first three dimensions. Another example would be that we have 3D points, and we want to find the closest points in one dimension (e.g. along the vertical axis of a trench).