Allow AdvectionCallback to stop a process before process time has elapsed (#34)
* Updated app readme. Added missing statement in examples CMake.

* Updated data source interface to allow for writing updated data.

* Updated psValueEstimator interface to use data instead of dataSource

* Changed name of volume model to advection callback.

This allows hooking into the advection process, which might be useful
for models other than those using cell sets.

* Fixed wrong function call in assert statement

* Added check to updateSurface() call in psProcess()

* PointLocators now use an axis mask instead of two different dimensions

This makes them more flexible, since we can say that only want the tree
to consider the x and z coordinate and ignore the y coordinate. This
change also enables the usecase for searching along only one axis (e.g.
to extract trench diameters at certain depths, by searching the closest
points along the vertical trench axis (point on the left sidewall and
point on the right sidewall) and use the distance between those two as a
measure for the diameter at that depth.

* Added comments to the point locator class and helper functions

* PointLocators now use an axis mask.

The axis mask allows selecting axes for nearest neighbors lookup. For
example: we could have a 4D data vector that consists of 3 spatial dimensions and 1 data dimension, so for the nearest neighbors lookup we would only want to use the first three dimensions. Another example would be that we have 3D points, and we want to find the closest points in one dimension (e.g. along the vertical axis of a trench).

* Exposed join function in CSVWriter. Minor include fixes.

* Added initial version of interpolation workflow example.

This example demonstrates the full workflow of generating data from
physical simulations, and using this data to reconstruct geometries by
interpolation. At the moment the main executable is still a work in
progress.

* Fixed bug in kNN interpolation, updated example.

* Major simplifications in psPointLocator design.

* Parameter and data updates.

* Fixed assert statement in GeometryReconstruction

* InterpolationWorkflow now uses different interpolation scheme for time

* Moved interpolation workflow to its own project

* Make verbose optional when installed

* KDTree now compiles without warnings.

* Unified KDTree point types.

* Marked some fields as private within node class

* KDTree now returns optional instead of nullptr if tree not initialized

* Value estimator now also uses optional return type for preventing ambiguous results

* Data workflow (data dimensions,...) is now fully dynamic.

* InterpolationDemo now uses the dimension provided in CSV file

* Reduced code duplication in Interpolation demo

* Untangled KDTree from rest of ViennaPS.

* Added plotting script to InterpolationDemo

* Using new prepareLS function in psProcess

* Added calls to prepareLS function to psProcess

* Cleanup of KDTree class

* Added nodiscard attribute to various function calls

* Removed unsupported method call

* Fixed formatting of example

* AdvectionCallback can now stop a running process early.